requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.18
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==35.0.0
flower==2.0.1
humanize==4.11.0
kombu==5.4.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
